perf(core/progress): `linear-progress` use svg for animations
Use svgs for smoother animations and less layout recalculations
