+++ packages/lib/src/components/Progress/index.ts
@@ -1,3 +1,2 @@
 export { default as CircularProgress } from './circular-progress.vue'
-export { default as LinearProgressSvg } from './linear-progress-svg.vue'
 export { default as LinearProgress } from './linear-progress.vue'
